be/requirements.txt
daphne==2.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
pytest==5.4.3
cryptography==2.9.2